+++ Molten.Math/SinglePrecision/Vectors/SByte4.cs
@@ -32,30 +32,30 @@ namespace Molten
 		///<summary>The size of <see cref="SByte4"/>, in bytes.</summary>
 		public static readonly int SizeInBytes = Marshal.SizeOf(typeof(SByte4));
 
-		///<summary>A SByte4 with every component set to 1.</summary>
-		public static readonly SByte4 One = new SByte4(1, 1, 1, 1);
+		///<summary>A SByte4 with every component set to (sbyte)1.</summary>
+		public static readonly SByte4 One = new SByte4((sbyte)1, (sbyte)1, (sbyte)1, (sbyte)1);
 
 		/// <summary>The X unit <see cref="SByte4"/>.</summary>
-		public static readonly SByte4 UnitX = new SByte4(1, 0, 0, 0);
+		public static readonly SByte4 UnitX = new SByte4((sbyte)1, (sbyte)0, (sbyte)0, (sbyte)0);
 
 		/// <summary>The Y unit <see cref="SByte4"/>.</summary>
-		public static readonly SByte4 UnitY = new SByte4(0, 1, 0, 0);
+		public static readonly SByte4 UnitY = new SByte4((sbyte)0, (sbyte)1, (sbyte)0, (sbyte)0);
 
 		/// <summary>The Z unit <see cref="SByte4"/>.</summary>
-		public static readonly SByte4 UnitZ = new SByte4(0, 0, 1, 0);
+		public static readonly SByte4 UnitZ = new SByte4((sbyte)0, (sbyte)0, (sbyte)1, (sbyte)0);
 
 		/// <summary>The W unit <see cref="SByte4"/>.</summary>
-		public static readonly SByte4 UnitW = new SByte4(0, 0, 0, 1);
+		public static readonly SByte4 UnitW = new SByte4((sbyte)0, (sbyte)0, (sbyte)0, (sbyte)1);
 
 		/// <summary>Represents a zero'd SByte4.</summary>
-		public static readonly SByte4 Zero = new SByte4(0, 0, 0, 0);
+		public static readonly SByte4 Zero = new SByte4((sbyte)0, (sbyte)0, (sbyte)0, (sbyte)0);
 
         /// <summary>
         /// Gets a value indicting whether this vector is zero
         /// </summary>
         public bool IsZero
         {
-            get => X == 0 && Y == 0 && Z == 0 && W == 0;
+            get => X == (sbyte)0 && Y == (sbyte)0 && Z == (sbyte)0 && W == (sbyte)0;
         }
 
 #region Constructors
